Ntr/5.7/merge 5.6

See merge request shopware/5/product/shopware!257

+++ engine/Shopware/Controllers/Frontend/Forms.php
@@ -160,6 +160,13 @@ class Shopware_Controllers_Frontend_Forms extends Enlight_Controller_Action
         /* @var Form $form */
         $form = $query->getOneOrNullResult(\Doctrine\ORM\AbstractQuery::HYDRATE_OBJECT);
 
+        if (!$form) {
+            throw new \Enlight_Controller_Exception(
+                'Form not found',
+                Enlight_Controller_Exception::Controller_Dispatcher_Controller_Not_Found
+            );
+        }
+
         if ($form->getAttribute()) {
             $this->getModelManager()->detach($form->getAttribute());
         }
@@ -167,10 +174,6 @@ class Shopware_Controllers_Frontend_Forms extends Enlight_Controller_Action
         $this->getModelManager()->detach($form);
         foreach ($form->getFields() as $field) {
             $this->getModelManager()->detach($field);
-        }
-
-        if (!$form) {
-            throw new \Enlight_Controller_Exception('Form not found', Enlight_Controller_Exception::Controller_Dispatcher_Controller_Not_Found);
         }
 
         /* @var Field $field */
@@ -390,8 +393,8 @@ class Shopware_Controllers_Frontend_Forms extends Enlight_Controller_Action
                     $placeholder1 = $placeholder;
                 }
 
-                $output .= "<input type=\"text\" class=\"{$element['class'][0]} $requiredField\" $requiredFieldAria value=\"{$post[0]}\" $placeholder0 id=\"{$element['name'][0]};{$element['name'][1]}\" name=\"{$element['name'][0]}\"/>\r\n";
-                $output .= "<input type=\"text\" class=\"{$element['class'][1]} $requiredField\" $requiredFieldAria value=\"{$post[1]}\" $placeholder1 id=\"{$element['name'][0]};{$element['name'][1]}\" name=\"{$element['name'][1]}\"/>\r\n";
+                $output .= "<input type=\"text\" class=\"{$element['class'][0]} $requiredField\" $requiredFieldAria value=\"{$post[0]}\" $placeholder0 id=\"{$element['name'][0]}\" name=\"{$element['name'][0]}\"/>\r\n";
+                $output .= "<input type=\"text\" class=\"{$element['class'][1]} $requiredField\" $requiredFieldAria value=\"{$post[1]}\" $placeholder1 id=\"{$element['name'][1]}\" name=\"{$element['name'][1]}\"/>\r\n";
                 break;
 
             case 'textarea':
